Solve Generate subsequence problem
Solved the problem to print all the possible subsequence of the list in python. 
Number of subsequence of the list with length N is 2^N. This solution uses recursion to create all the sub sequence. The approach is a decision tree where at each position, the decision needs to be made whether to include a number from list or not.